Improve windows support
It seems the unzip binary that comes with Git for Windows is slightly different from unzip on Linux and macOS. While I was using this `windows-latest`, this script would cause unzip to _only_ extract the files that are directly in the `go/pkg` and `go/src` dirs causing the sources of all built-in packages to be unavailable. This changes the call to recursively extract the files for those two dirs.

In addition `MSWin32` was added to the test for downloading the win binaries. This allows for the script to be run directly from PowerShell, at least on GitHub actions. It'll likely work everywhere as long as the curl/unzip (and other dependency) binaries are available in the PATH.